ufro-master/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx[http2]==0.27.2

motor==3.6.0
pymongo==4.9.1
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
pyyaml==6.0.2
python-multipart==0.0.9

gunicorn==23.0.0
